refactor `model` and `rhcos` to artcommonlib
