+++ art-common/artcommonlib/rhcos.py
@@ -1,0 +1,77 @@
+from artcommonlib.model import ListModel, Model
+
+# Historically the only RHCOS container was 'machine-os-content'; see
+# https://github.com/openshift/machine-config-operator/blob/master/docs/OSUpgrades.md
+# But with OCP 4.12 this changed, see
+# https://github.com/coreos/enhancements/blob/main/os/coreos-layering.md
+default_primary_container = dict(
+    name="machine-os-content",
+    build_metadata_key="oscontainer",
+    primary=True)
+
+
+class RhcosMissingContainerException(Exception):
+    """
+    Thrown when group.yml configuration expects an RHCOS container but it is
+    not available as specified in the RHCOS metadata.
+    """
+    pass
+
+
+def get_container_configs(runtime):
+    """
+    look up the group.yml configuration for RHCOS container(s) for this group, or create if missing.
+    @return ListModel with Model entries like ^^ default_primary_container
+    """
+    return runtime.group_config.rhcos.payload_tags or ListModel([default_primary_container])
+
+
+def get_container_names(runtime):
+    """
+    look up the payload tags of the group.yml-configured RHCOS container(s) for this group
+    @return list of container names
+    """
+    return {tag.name for tag in get_container_configs(runtime)}
+
+
+def get_primary_container_conf(runtime):
+    """
+    look up the group.yml-configured primary RHCOS container for this group.
+    @return Model with entries for name and build_metadata_key
+    """
+    for tag in get_container_configs(runtime):
+        if tag.primary:
+            return tag
+    raise Exception("Need to provide a group.yml rhcos.payload_tags entry with primary=true")
+
+
+def get_primary_container_name(runtime):
+    """
+    convenience method to retrieve configured primary RHCOS container name
+    @return primary container name (used in payload tag)
+    """
+    return get_primary_container_conf(runtime).name
+
+
+def get_container_pullspec(build_meta: dict, container_conf: Model) -> str:
+    """
+    determine the container pullspec from the RHCOS build meta and config
+    @return full container pullspec string (registry/repo@sha256:...)
+    """
+    key = container_conf.build_metadata_key
+    if key not in build_meta:
+        raise RhcosMissingContainerException(f"RHCOS build {build_meta['buildid']} has no '{key}' attribute in its metadata")
+
+    container = build_meta[key]
+
+    if 'digest' in container:
+        # "oscontainer": {
+        #   "digest": "sha256:04b54950ce2...",
+        #   "image": "quay.io/openshift-release-dev/ocp-v4.0-art-dev"
+        # },
+        return container['image'] + "@" + container['digest']
+
+    # "base-oscontainer": {
+    #     "image": "registry.ci.openshift.org/rhcos/rhel-coreos@sha256:b8e1064cae637f..."
+    # },
+    return container['image']
