+++ art-common/pyproject.toml
@@ -1,24 +1,22 @@
 [build-system]
-requires = ["setuptools>=65.5.1", "setuptools_scm[toml]>=8.0"]
+requires = [
+    "setuptools>=65.5.1",  # not directly required, pinned by Snyk to avoid a vulnerability
+]
 build-backend = "setuptools.build_meta"
 
-[tool.setuptools.packages.find]
-include = ["elliottlib*"]
-namespaces = false
-
-[tool.setuptools_scm]
-root = ".."
+[tool.setuptools]
+packages = ["artcommonlib"]
 
 [project]
 dynamic = ["version"]
-name="rh-elliott"
+name="rh-art-commonlib"
 authors = [
     {name = "AOS ART Team", email = "aos-team-art@redhat.com"},
 ]
 description = "CLI tool for managing and automating Red Hat software releases"
 readme = "README.md"
 requires-python = ">=3.8"
-keywords = ["openshift", "errata", "advisory"]
+keywords = ["openshift"]
 license.file = "LICENSE"
 classifiers = [
     "Development Status :: 5 - Production/Stable",
@@ -42,40 +40,16 @@ classifiers = [
     "Natural Language :: English",
 ]
 dependencies = [
-    "aiohttp[speedups] >= 3.6",
-    "click >= 8.1.3",
-    "errata-tool ~= 1.31.0",
-    "future",
-    "koji >= 1.18",
-    "semver",
     "pip_system_certs",
-    "pydantic ~= 1.10.7",
-    "pygit2 == 1.10.1",  # https://github.com/libgit2/pygit2/issues/1176
-    "python-bugzilla >= 3.2",
-    "pyyaml",
-    "requests",
-    "requests-gssapi ~= 1.2.3",
-    "rh-art-commonlib",
-    "ruamel.yaml",
-    "tenacity",
-    "jira >= 3.4.1",  # https://github.com/pycontribs/jira/issues/1486
 ]
 
-[project.scripts]
-elliott = "elliottlib.cli.__main__:main"
-
 [project.urls]
-homepage = "https://github.com/openshift-eng/art-tools/tree/main/elliott"
+homepage = "https://github.com/openshift-eng/art-tools"
 repository = "https://github.com/openshift-eng/art-tools.git"
 
 [project.optional-dependencies]
 tests = [
-    # test reqs
-    "flexmock",
-    # tox reqs
-    "coverage>=5.3",
-    "flake8>=3.8.4",
-    "tox>=3.20.1",
+    # test reqs - well, no tests yet
     # makefile reqs
     "virtualenv>=20.2.2",
     "pytest>=6.2.1",
