+++ doozer/doozerlib/assembly_inspector.py
@@ -10,7 +10,8 @@ from doozerlib.image import BrewBuildImageInspector
 from doozerlib.rpmcfg import RPMMetadata
 from doozerlib.assembly import assembly_rhcos_config, AssemblyTypes, assembly_permits, AssemblyIssue, \
     AssemblyIssueCode, assembly_type
-from doozerlib.rhcos import RHCOSBuildInspector, RHCOSBuildFinder, get_container_configs, RhcosMissingContainerException
+from doozerlib.rhcos import RHCOSBuildInspector, RHCOSBuildFinder
+from artcommonlib.rhcos import get_container_configs, RhcosMissingContainerException
 
 
 class AssemblyInspector:
