+++ doozer/requirements.txt
@@ -22,3 +22,4 @@ setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerabil
 aiohttp
 jira==3.4.1
 ghapi
+rh-art-commonlib
